play notebook based on DocVectors
This notebook was put together to explore an LSTM variation of Minmen Chen 2017's Document Vector work. However, the Document Vector was not sufficient to predict sentiment. Brighid is putting this notebook up on this project for now in case anyone else wants to work it or play around with it.